modeler API (changes due to review) : mostly rename directory from "optim" to "modeler"
