+++ src/solver/modeler/api/include/antares/solver/modeler/api/linearProblem.h
