+++ src/solver/modeler/api/include/antares/solver/modeler/api/linearProblemData.h
