+++ src/solver/modeler/ortoolsImpl/mipConstraint.cpp
@@ -22,8 +22,8 @@
 #include <ortools/linear_solver/linear_solver.h>
 
 #include <antares/logs/logs.h>
-#include <antares/solver/optim/ortoolsImpl/mipConstraint.h>
-#include <antares/solver/optim/ortoolsImpl/mipVariable.h>
+#include <antares/solver/modeler/ortoolsImpl/mipConstraint.h>
+#include <antares/solver/modeler/ortoolsImpl/mipVariable.h>
 
 namespace Antares::Solver::Optim::OrtoolsImpl
 {
