+++ src/tests/src/solver/CMakeLists.txt
@@ -4,4 +4,4 @@ add_subdirectory(optimisation)
 add_subdirectory(utils)
 add_subdirectory(infeasible-problem-analysis)
 add_subdirectory(lps)
-add_subdirectory(optim)
+add_subdirectory(modeler)
