+++ src/tests/src/solver/modeler/api/testApiOrtoolsLinearProblem.cpp
@@ -18,13 +18,13 @@
  * You should have received a copy of the Mozilla Public Licence 2.0
  * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
  */
-#define BOOST_TEST_MODULE test optim api ortools
+#define BOOST_TEST_MODULE test modeler api ortools
 
 #define WIN32_LEAN_AND_MEAN
 
 #include <boost/test/unit_test.hpp>
 
-#include <antares/solver/optim/ortoolsImpl/linearProblem.h>
+#include <antares/solver/modeler/ortoolsImpl/linearProblem.h>
 
 using namespace Antares::Solver::Optim;
